Wizard: Tweaked login dialog.
The login dialog was not showing up, since the jquery.reveal script
needs to be included at the ending of the html body.

Clicking on a slide > 0 without associating a base account will show
the association dialog for those forcefully.

Other details of the login process were tweaked, and a lot of the wizard
code was cleand up.